requirements.txt
alembic==1.18.4
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
gunicorn==25.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
SQLAlchemy==2.0.47
typing_extensions==4.15.0
Werkzeug==3.1.6
